GenesisState: move logic out of constructor
